requirement.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
opencv-python>=4.8.1.78
ultralytics>=8.1.0
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.3
pillow>=10.1.0
python-dotenv>=1.0.0
websockets>=12.0
